forecasting/Chronos+TiDE/requirements.txt
git+https://github.com/amazon-science/chronos-forecasting.git
u8darts[all]==0.30.0
tensorflow==2.16.1
torch==2.3.1
scikit-learn==1.5.0
pandas==2.2.2
dask[dataframe]==2024.6.2
seaborn==0.13.2
tqdm==4.66.4
jupyter==1.0.0
numpy==1.26.4
datasets==2.20.0